--- Assets/game/level1/gate.meta
@@ -1,10 +1,0 @@
-fileFormatVersion: 2
-guid: 119187cc784913c41a80766a4913087e
-folderAsset: yes
-timeCreated: 1607620834
-licenseType: Free
-DefaultImporter:
-  externalObjects: {}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
