+++ Assets/game/level1/sensor.prefab.meta
