--- Assets/pau1234/UI/Text.controller.meta
@@ -1,10 +1,0 @@
-fileFormatVersion: 2
-guid: 83b3dc06952c5fb419f382b374dfb267
-timeCreated: 1607911522
-licenseType: Free
-NativeFormatImporter:
-  externalObjects: {}
-  mainObjectFileID: 9100000
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
