--- Assets/pau1234/UI/back.anim.meta
@@ -1,10 +1,0 @@
-fileFormatVersion: 2
-guid: d38aa2682e211714fa9b0c0b0c805e55
-timeCreated: 1607878691
-licenseType: Free
-NativeFormatImporter:
-  externalObjects: {}
-  mainObjectFileID: 7400000
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
