+++ Assets/pau1234/animate.meta
@@ -1,10 +1,10 @@
 fileFormatVersion: 2
-guid: ee27e09713e54ff4384a0c408bf0c996
-timeCreated: 1606543464
+guid: 460f15c4cdad3e540af67eb0ca06a675
+folderAsset: yes
+timeCreated: 1608103869
 licenseType: Free
-NativeFormatImporter:
+DefaultImporter:
   externalObjects: {}
-  mainObjectFileID: 100100000
   userData: 
   assetBundleName: 
   assetBundleVariant:
