+++ Assets/pau1234/map/gate.cs.meta
