+++ Assets/pau1234/map_empty_prefab/gate/close.controller.meta
@@ -1,10 +1,10 @@
 fileFormatVersion: 2
-guid: 89a0e5c377dc9504b80d5d55a39a87d5
-timeCreated: 1607942310
+guid: 53b7480c79e35e744a272f7cbfde7e90
+timeCreated: 1608166802
 licenseType: Free
 NativeFormatImporter:
   externalObjects: {}
-  mainObjectFileID: 7400000
+  mainObjectFileID: 9100000
   userData: 
   assetBundleName: 
   assetBundleVariant:
