+++ Assets/pau1234/map_empty_prefab/gate/open.anim.meta
@@ -1,10 +1,10 @@
 fileFormatVersion: 2
-guid: 05feb91af1a613f4fb8d03255643805b
-timeCreated: 1606543484
+guid: 8dbf28d76d182ee4aa4959c3fd6f7ff3
+timeCreated: 1608166945
 licenseType: Free
 NativeFormatImporter:
   externalObjects: {}
-  mainObjectFileID: 100100000
+  mainObjectFileID: 7400000
   userData: 
   assetBundleName: 
   assetBundleVariant:
