+++ Assets/resource/room1_1.asset
@@ -12,4 +12,4 @@ MonoBehaviour:
   m_Name: room1_1
   m_EditorClassIdentifier: 
   roomMap: {fileID: 4900000, guid: 9d54f8155138cbe47ad0d077e6579e54, type: 3}
-  monsterWave: 0a0000000a00000001000000
+  monsterWave: 01000000
