+++ Assets/scene/test.unity
@@ -294,10 +294,6 @@ Prefab:
       propertyPath: m_Pivot.y
       value: 0
       objectReference: {fileID: 0}
-    - target: {fileID: 1739754512929474, guid: 64ca5b4723fb9f641951acc237508d68, type: 2}
-      propertyPath: m_Name
-      value: UI
-      objectReference: {fileID: 0}
     m_RemovedComponents: []
   m_ParentPrefab: {fileID: 100100000, guid: 64ca5b4723fb9f641951acc237508d68, type: 2}
   m_IsPrefabParent: 0
